Address review: payload migration seam, split graph registry from url_utils, drop no-op commit, document scroll delays
